fix: Choose the correct protocol
During development, we use the HTTP protocol. In production, we use the HTTPS protocol.

As a result, we must strictly choose a protocol for development and production builds.

Now we can use both HTTP and HTTPS protocols, but this is incorrect. This is because https://stage.creaty.club and http://creaty.club (after merge) will not work.